+++ login-workflow/.husky/pre-commit
@@ -1,0 +1,4 @@
+#!/usr/bin/env sh
+. "$(dirname -- "$0")/_/husky.sh"
+
+cd ./login-workflow && yarn prettier:check && yarn lint && yarn test
