+++ login-workflow/example/src/navigation/navigation-drawer.tsx
@@ -25,7 +25,7 @@ export const navGroupItems: NavItem[] = [
         itemID: 'RegistratonContextExample',
         icon: { name: 'person' },
     },
-    {        
+    {
         title: 'Registration Provider',
         itemID: 'RegistrationProviderExample',
         icon: { name: 'app-registration' },
