+++ login-workflow/src/components/SetPassword/SetPassword.tsx
@@ -100,7 +100,7 @@ export const SetPassword: React.FC<React.PropsWithChildren<SetPasswordProps>> = 
         }
 
         return true;
-    }, [passwordRequirements, passwordInput]);
+    }, [passwordRequirements, t, passwordInput]);
     return (
         <>
             {children}
